pyproject.toml
[tool.poetry]
name = "pycziutils"
version = "0.1.0"
description = "Python utilities to read CZI files and parse metadata through python-bioformats"
authors = ["Yohsuke T. Fukai <ysk@yfukai.net>"]
license = "BSD license"

readme = "README.rst"

repository = "https://github.com/yfukai/pycziutils"

homepage = "https://pycziutils.readthedocs.io"


packages = [
    { include = "pycziutils", from = "src"},
]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]


[tool.poetry.dependencies]
python = "^3.7.1"
python-bioformats = "^4.0.4"
python-javabridge = "^4.0.3"
xmltodict = "^0.12"
numpy = "^1.9"
pandas = "^1.0"

[tool.poetry.dev-dependencies]
bump2version = "^1"
pip = ">=18.1"
wheel = "^0.32.1"
watchdog = "^2"
flake8 = "^3.5"
tox = "^3.5.2"
coverage = "^5"
Sphinx = "^3.5.3"
sphinx-rtd-theme = "^0.5.2"
twine = "^3"

pytest = "^6"
pytest-runner = ">=4.2"
pytest-cov = ">=2.6.0"
setuptools = ">=40.5.0"
rstcheck = "^3.3.1"
ptpython = ">=2.0.4"
pysen = {version = "0.9.1", extras = ["lint"]}
pysen-ls = "^0.1.0"
pytest-datadir = "^1.3.1"

[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
#mypy_preset = "strict"
line_length = 88
py_version = "py37"

[tool.pysen.lint.source]
  excludes = ["docs/"]